fix missing dependencies under npm@3
The scenario we need to test for is npm3's flattening behaviour, which
results in our 'cached' modules not containing any of their dependencies
like they do with npm@2. The use of cowsay here is because mkdirp is a
dependency of mocha, so both mkdirp and mkdirp's dependencies are
present in the module resolution path when we run the sandbox test. The
cowsay module depends on modules that are not used in the rest of
strong-cached-install, so there is no way for strong-cached-install's
dependencies to mask the problem.

By adding the --global-style flag we tell npm to treat 'npm install foo'
similar to a global install in that it installs all its dependencies
under foo's node_modules, which is what we really want for caching.

We also switch from relying on $PATH to referencing the .bin directory
directly because it removes a potential source of false positives from
the test command already being in the $PATH for unrelated reasons.